+++ src/main/java/com/my/appWordle/controllers/MatchController.java
@@ -3,6 +3,8 @@ package com.my.appWordle.controllers;
 import com.my.appWordle.models.Match;
 import com.my.appWordle.repositories.MatchRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,29 +16,47 @@ public class MatchController {
     private MatchRepository matchRepository;
 
     @GetMapping
-    public List<Match> getAllMatches() {
-        return matchRepository.findAll();
+    public ResponseEntity<List<Match>> getAllMatches() {
+        List<Match> matches = matchRepository.findAll();
+
+        if (matches.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>(matches, HttpStatus.OK);
+        }
     }
 
     @GetMapping("/{idMatch}")
-    public Match getMatchById(@PathVariable Long idMatch) {
-        return matchRepository.findById(idMatch).orElse(null);
+    public ResponseEntity<Match> getMatchById(@PathVariable Long idMatch) {
+        Match match = matchRepository.findById(idMatch).orElse(null);
+
+        if (match != null) {
+            return new ResponseEntity<>(match, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping
-    public Match createMatch(@RequestBody Match match) {
-        return matchRepository.save(match);
+    public ResponseEntity<Match> createMatch(@RequestBody Match match) {
+        Match createdMatch = matchRepository.save(match);
+        return new ResponseEntity<>(createdMatch, HttpStatus.CREATED);
     }
 
     @PutMapping("/{idMatch}")
-    public Match updateMatch(@PathVariable Long idMatch, @RequestBody Match match) {
+    public ResponseEntity<Match> updateMatch(@PathVariable Long idMatch, @RequestBody Match match) {
         match.setIdMatch(idMatch);
-        return matchRepository.save(match);
+        Match updatedMatch = matchRepository.save(match);
+        return new ResponseEntity<>(updatedMatch, HttpStatus.OK);
     }
 
     @DeleteMapping("/{idMatch}")
-    public void deleteMatch(@PathVariable Long idMatch) {
-        matchRepository.deleteById(idMatch);
+    public ResponseEntity<Void> deleteMatch(@PathVariable Long idMatch) {
+        if (matchRepository.existsById(idMatch)) {
+            matchRepository.deleteById(idMatch);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
-
