+++ src/main/java/com/my/appWordle/controllers/TeamController.java
@@ -3,6 +3,8 @@ package com.my.appWordle.controllers;
 import com.my.appWordle.models.Team;
 import com.my.appWordle.repositories.TeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,29 +16,47 @@ public class TeamController {
     private TeamRepository teamRepository;
 
     @GetMapping
-    public List<Team> getAllTeams() {
-        return teamRepository.findAll();
+    public ResponseEntity<List<Team>> getAllTeams() {
+        List<Team> teams = teamRepository.findAll();
+
+        if (teams.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>(teams, HttpStatus.OK);
+        }
     }
 
     @GetMapping("/{idTeam}")
-    public Team getTeamById(@PathVariable Long idTeam) {
-        return teamRepository.findById(idTeam).orElse(null);
+    public ResponseEntity<Team> getTeamById(@PathVariable Long idTeam) {
+        Team team = teamRepository.findById(idTeam).orElse(null);
+
+        if (team != null) {
+            return new ResponseEntity<>(team, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping
-    public Team createTeam(@RequestBody Team team) {
-        return teamRepository.save(team);
+    public ResponseEntity<Team> createTeam(@RequestBody Team team) {
+        Team createdTeam = teamRepository.save(team);
+        return new ResponseEntity<>(createdTeam, HttpStatus.CREATED);
     }
 
     @PutMapping("/{idTeam}")
-    public Team updateTeam(@PathVariable Long idTeam, @RequestBody Team team) {
+    public ResponseEntity<Team> updateTeam(@PathVariable Long idTeam, @RequestBody Team team) {
         team.setIdTeam(idTeam);
-        return teamRepository.save(team);
+        Team updatedTeam = teamRepository.save(team);
+        return new ResponseEntity<>(updatedTeam, HttpStatus.OK);
     }
 
     @DeleteMapping("/{idTeam}")
-    public void deleteTeam(@PathVariable Long idTeam) {
-        teamRepository.deleteById(idTeam);
+    public ResponseEntity<Void> deleteTeam(@PathVariable Long idTeam) {
+        if (teamRepository.existsById(idTeam)) {
+            teamRepository.deleteById(idTeam);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
-
